Issue #402 - Move profiles check to workflowproj module (#403)
